requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
animec==0.3.3
Async-OAuthlib==0.0.9
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.4.0
asyncprawcore==2.3.0
attrs==21.2.0
beautifulsoup4==4.10.0
discord-components==2.1.1
discord-ext-ipc==2.1.1
discord.py==1.7.3
DiscordUtils==1.3.4
ffmpeg==1.4
google-api-core==2.0.1
google-api-python-client==2.21.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
idna==3.2
import-expression==1.1.4
ipc==1.0.1
Jinja2==3.0.1
jishaku==2.3.0
oauthlib==3.1.1
opencv-python==4.5.3.56
packaging==21.0
pathy==0.6.0
Pillow==8.3.2
Pint==0.17
pipenv==2018.11.26
py-session==0.1.1
pyasn1==0.4.8
asyncpg
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pyfiglet==0.8.post1
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.7.5
Quart==0.15.1
Quart-Discord==2.1.4
typing-extensions==3.10.0.2
urllib3==1.26.6
wikipedia==1.4.0
yarl==1.6.3
youtube-dl==2021.6.6
discord-py-slash-command
akinator.py
prsaw
requests
psutil
timeago
jikanpy
googletrans
